requirements.txt
asgiref==3.5.0
astroid==2.15.6
async-timeout==4.0.2
dill==0.3.7
dj-database-url==0.5.0
Django==4.0.10
django-cors-headers==3.11.0
django-debug-toolbar==4.1.0
django-filter==23.2
django-heroku==0.3.1
django-mssql==1.8
django-mssql-backend==2.8.1
django-pymssql==1.7.1
django-redis==5.3.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mssql-django==1.3
platformdirs==3.10.0
psycopg2==2.9.6
PyJWT==2.3.0
pylint==2.17.5
pymssql==2.2.7
pyodbc==4.0.39
pypyodbc==1.3.6
python-dotenv==1.0.0
pytz==2021.3
redis==4.5.5
simplejson==3.17.6
sqlparse==0.4.2
tomlkit==0.12.1
tzdata==2021.5
whitenoise==6.1.0
wrapt==1.15.0
